some moving around, creating pkgs and unit test WIP
