+++ pytrader/logger.py
@@ -4,10 +4,10 @@ from datetime import datetime
 
 logging.config.dictConfig({'version': 1, 'disable_existing_loggers': True,})
 
-formatter = logging.Formatter('%(asctime)s.%(msecs)03d %(levelname)s %(message)s', '%Y-%m-%d_%I-%M-%S')
+formatter = logging.Formatter('%(asctime)s.%(msecs)03d %(levelname)s %(message)s', '%Y-%m-%d_%H-%M-%S')
 
 def config_logger(level=logging.INFO):
-    handler = logging.FileHandler(filename=f"pytrader/logs/{datetime.now():%Y-%m-%d_%I-%M-%S}.log")        
+    handler = logging.FileHandler(filename=f"pytrader/logs/{datetime.now():%Y-%m-%d_%H-%M-%S}.log")        
     handler.setFormatter(formatter)
 
     logger = logging.getLogger()
